refactor: Post의 조회수와 관련된 인수 테스트를 별도의 클래스로 분리
