+++ backend/src/test/java/edonymyeon/backend/post/integration/PostIntegrationTest.java
@@ -34,7 +34,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
 @SuppressWarnings("NonAsciiCharacters")
-public class PostIntegrationTest extends IntegrationTest implements ImageFileCleaner {
+public class PostIntegrationTest extends IntegrationTest  implements ImageFileCleaner {
 
     private static File 이미지1 = new File("./src/test/resources/static/img/file/test_image_1.jpg");
     private static File 이미지2 = new File("./src/test/resources/static/img/file/test_image_2.jpg");
@@ -658,37 +658,6 @@ public class PostIntegrationTest extends IntegrationTest implements ImageFileCle
                             .isEqualTo(POST_IMAGE_COUNT_INVALID.getMessage());
                 }
         );
-    }
-
-    @Test
-    void 게시글_조회수_시나리오() {
-        final Member 작성자 = this.사용자를_하나_만든다();
-        final Member 열람자 = this.사용자를_하나_만든다();
-        final Post 게시글 = this.postTestSupport.builder()
-                .member(작성자)
-                .build();
-
-        작성자가_자신의_글을_조회하면_조회수가_오르지_않는다(작성자, 게시글);
-        for (int count = 1; count <= 4; count++) {
-            다른사람이_글을_조회하면_조회수가_오른다(열람자, 게시글, count);
-        }
-    }
-
-    private void 다른사람이_글을_조회하면_조회수가_오른다(final Member 열람자, final Post 게시글, final int expected) {
-        final ExtractableResponse<Response> 다른사람이_게시글_조회한_결과 = this.게시글_하나를_상세_조회한다(열람자, 게시글.getId());
-        System.out.println(다른사람이_게시글_조회한_결과.body().asPrettyString());
-        assertThat(다른사람이_게시글_조회한_결과.body()
-                .jsonPath()
-                .getLong("reactionCount.viewCount"))
-                .isEqualTo(expected);
-    }
-
-    private void 작성자가_자신의_글을_조회하면_조회수가_오르지_않는다(final Member 작성자, final Post 게시글) {
-        다른사람이_글을_조회하면_조회수가_오른다(작성자, 게시글, 0);
-    }
-
-    public static void main(String[] args) {
-        System.out.println(new Scanner(System.in).nextLine().replaceAll(" ", "_"));
     }
 
     private ExtractableResponse<Response> 게시글_하나를_상세_조회한다(final Member 열람인, final long 게시글_id) {
